fix  de score suma.

+++ ProyectoEclipse/src/Usuario.java
@@ -43,7 +43,7 @@ public class Usuario  {
 	// Para que el score este aca, el usuario tiene que conocer a todos los juegos. Esto genera dependencia.
 	// Para mi lo mejor es pasar el user al controlador del juego y que el score lo tenga el. 
 	
-	public void sumarScore(float puntos){
+	public void sumarScore(int puntos){
 		this.score = puntos + this.score;
 	}
 }
